add TI-TieBreakingOpenList (move template class definition from .cc to .h for Best Practice).
